配表，协程追踪，控制台

+++ Assets/_Project/Scripts/Systems/MarketEventSystem.cs
@@ -133,7 +133,7 @@ public class MarketEventSystem : SingletonCom<MarketEventSystem>
             return;
         }
 
-        var ga = new ChangeStockPriceGA(cv, stockType, null, percentDict);
+        var ga = new ChangeStockPriceGA(cv, (int)stockType, null, percentDict);
         ActionSystem.Ins.Perform(ga);
     }
 }
